+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Note that since we don't clearly distinguish between a public and private interf
 ## [Unreleased]
 
 - Fix polymer trace shift for PNA
+- Use carbon for all trace atoms (to play nicely with color themes)
 
 ## [v3.32.0] - 2023-03-20